Make text layout an optional feature
This permits trimming dependencies
