+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "femtovg"
 description = "Antialiased 2D vector drawing library"
-version = "0.18.1"
+version = "0.19.0"
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 authors = [
@@ -19,22 +19,23 @@ rust-version = "1.85"
 debug = true
 
 [dependencies]
-fnv = "1.0.7"
+fnv = { version = "1.0.7" }
 rgb = "0.8.50"
 bytemuck = { version = "1.16", features = ["derive"] }
 imgref = "1.11.0"
 bitflags = "2.6.0"
-rustybuzz = "0.20.0"
-unicode-bidi = "0.3.17"
-unicode-segmentation = "1.12.0"
+rustybuzz = { version = "0.20.0", optional = true }
+unicode-bidi = { version = "0.3.17", optional = true }
+unicode-segmentation = { version = "1.12.0", optional = true }
 slotmap = "1.0.7"
-lru = { version = "0.16.0", default-features = false }
+lru = { version = "0.16.0", optional = true, default-features = false }
 image = { version = "0.25.0", optional = true, default-features = false }
 serde = { version = "1.0", optional = true, features = ["derive", "rc"] }
 glow = { version = "0.16.0", default-features = false }
 log = "0.4"
 wgpu = { version = "26", optional = true, default-features = false, features = ["wgsl"] }
 itertools = "0.14"
+ttf-parser = { version = "0.25.1" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 glutin = { version = "0.31", optional = true, default-features = false }
@@ -48,10 +49,11 @@ web_sys = { version = "0.3", package = "web-sys", features = [
 wasm-bindgen = "0.2"
 
 [features]
-default = ["image-loading"]
+default = ["image-loading", "textlayout"]
 image-loading = ["image"]
 debug_inspector = []
 wgpu = ["dep:wgpu"]
+textlayout = ["dep:rustybuzz", "dep:unicode-bidi", "dep:unicode-segmentation", "dep:lru"]
 
 [dev-dependencies]
 winit = { version = "0.29.1" }
