+++ src/text.rs
@@ -1,25 +1,12 @@
-use std::{
-    borrow::Borrow,
-    cell::RefCell,
-    convert::TryInto,
-    ffi::OsStr,
-    fs,
-    hash::{Hash, Hasher},
-    ops::Range,
-    path::Path as FilePath,
-    rc::Rc,
-};
+use std::{borrow::Borrow, cell::RefCell, ffi::OsStr, fs, hash::Hash, path::Path as FilePath, rc::Rc};
 
-use fnv::{FnvBuildHasher, FnvHashMap, FnvHasher};
-use lru::LruCache;
+use fnv::FnvHashMap;
+#[cfg(feature = "textlayout")]
 use rustybuzz::ttf_parser;
 use slotmap::{DefaultKey, SlotMap};
 
-use unicode_bidi::BidiInfo;
-use unicode_segmentation::UnicodeSegmentation;
-
 use crate::{
-    paint::{PaintFlavor, StrokeSettings, TextSettings},
+    paint::{PaintFlavor, StrokeSettings},
     Canvas, Color, ErrorKind, FillRule, ImageFlags, ImageId, ImageInfo, Paint, PixelFormat, PositionedGlyph,
     RenderTarget, Renderer,
 };
@@ -31,6 +18,11 @@ mod font;
 pub use font::FontMetrics;
 use font::{Font, GlyphRendering};
 
+#[cfg(feature = "textlayout")]
+mod textlayout;
+#[cfg(feature = "textlayout")]
+pub use textlayout::*;
+
 // This padding is an empty border around the glyph’s pixels but inside the
 // sampled area (texture coordinates) for the quad in render_atlas().
 const GLYPH_PADDING: u32 = 1;
@@ -41,6 +33,7 @@ const GLYPH_PADDING: u32 = 1;
 const GLYPH_MARGIN: u32 = 1;
 
 const TEXTURE_SIZE: usize = 512;
+#[cfg(feature = "textlayout")]
 const DEFAULT_LRU_CACHE_CAPACITY: usize = 1000;
 
 /// A font handle.
@@ -132,56 +125,6 @@ pub struct RenderedGlyph {
     color_glyph: bool,
 }
 
-#[derive(Copy, Clone, Debug)]
-pub struct ShapedGlyph {
-    pub x: f32,
-    pub y: f32,
-    pub c: char,
-    pub byte_index: usize,
-    pub font_id: FontId,
-    pub glyph_id: u16,
-    pub width: f32,
-    pub height: f32,
-    pub advance_x: f32,
-    pub advance_y: f32,
-    pub offset_x: f32,
-    pub offset_y: f32,
-    pub bearing_x: f32,
-    pub bearing_y: f32,
-}
-
-#[derive(Clone, Debug, Default)]
-struct ShapedWord {
-    glyphs: Vec<ShapedGlyph>,
-    width: f32,
-}
-
-#[derive(Copy, Clone, Hash, Eq, PartialEq, Ord, PartialOrd)]
-struct ShapingId {
-    size: u32,
-    word_hash: u64,
-    font_ids: [Option<FontId>; 8],
-}
-
-impl ShapingId {
-    fn new(font_size: f32, font_ids: [Option<FontId>; 8], word: &str, max_width: Option<f32>) -> Self {
-        let mut hasher = FnvHasher::default();
-        word.hash(&mut hasher);
-        if let Some(max_width) = max_width {
-            (max_width.trunc() as i32).hash(&mut hasher);
-        }
-
-        Self {
-            size: (font_size * 10.0).trunc() as u32,
-            word_hash: hasher.finish(),
-            font_ids,
-        }
-    }
-}
-
-type ShapedWordsCache<H> = LruCache<ShapingId, Result<ShapedWord, ErrorKind>, H>;
-type ShapingRunCache<H> = LruCache<ShapingId, TextMetrics, H>;
-
 pub struct FontTexture {
     pub atlas: Atlas,
     pub(crate) image_id: ImageId,
@@ -237,73 +180,38 @@ impl TextContext {
         self.0.borrow_mut().add_shared_font_with_index(data, face_index)
     }
 
-    /// Returns information on how the provided text will be drawn with the specified paint.
-    pub fn measure_text<S: AsRef<str>>(
-        &self,
-        x: f32,
-        y: f32,
-        text: S,
-        paint: &Paint,
-    ) -> Result<TextMetrics, ErrorKind> {
-        self.0.borrow_mut().measure_text(x, y, text, &paint.text)
-    }
-
-    /// Returns the maximum index-th byte of text that will fit inside `max_width`.
-    ///
-    /// The retuned index will always lie at the start and/or end of a UTF-8 code point sequence or at the start or end of the text
-    pub fn break_text<S: AsRef<str>>(&self, max_width: f32, text: S, paint: &Paint) -> Result<usize, ErrorKind> {
-        self.0.borrow_mut().break_text(max_width, text, &paint.text)
-    }
-
-    /// Returnes a list of ranges representing each line of text that will fit inside `max_width`
-    pub fn break_text_vec<S: AsRef<str>>(
-        &self,
-        max_width: f32,
-        text: S,
-        paint: &Paint,
-    ) -> Result<Vec<Range<usize>>, ErrorKind> {
-        self.0.borrow_mut().break_text_vec(max_width, text, &paint.text)
-    }
-
     /// Returns font metrics for a particular Paint.
     pub fn measure_font(&self, paint: &Paint) -> Result<FontMetrics, ErrorKind> {
         self.0
             .borrow_mut()
             .measure_font(paint.text.font_size, &paint.text.font_ids)
     }
-
-    /// Adjusts the capacity of the shaping run cache. This is a cache for measurements of whole
-    /// strings.
-    pub fn resize_shaping_run_cache(&self, capacity: std::num::NonZeroUsize) {
-        self.0.borrow_mut().resize_shaping_run_cache(capacity)
-    }
-
-    /// Adjusts the capacity of the shaped words cache. This is a cache for measurements of
-    /// individual words. Words are separated by
-    /// [UAX#29 word boundaries](http://www.unicode.org/reports/tr29/#Word_Boundaries).
-    pub fn resize_shaped_words_cache(&self, capacity: std::num::NonZeroUsize) {
-        self.0.borrow_mut().resize_shaped_words_cache(capacity)
-    }
 }
 
 pub struct TextContextImpl {
     fonts: SlotMap<DefaultKey, Font>,
-    shaping_run_cache: ShapingRunCache<FnvBuildHasher>,
-    shaped_words_cache: ShapedWordsCache<FnvBuildHasher>,
+    #[cfg(feature = "textlayout")]
+    shaping_run_cache: textlayout::ShapingRunCache<fnv::FnvBuildHasher>,
+    #[cfg(feature = "textlayout")]
+    shaped_words_cache: textlayout::ShapedWordsCache<fnv::FnvBuildHasher>,
 }
 
 impl Default for TextContextImpl {
     fn default() -> Self {
-        let fnv_run = FnvBuildHasher::default();
-        let fnv_words = FnvBuildHasher::default();
+        #[cfg(feature = "textlayout")]
+        let fnv_run = fnv::FnvBuildHasher::default();
+        #[cfg(feature = "textlayout")]
+        let fnv_words = fnv::FnvBuildHasher::default();
 
         Self {
             fonts: SlotMap::default(),
-            shaping_run_cache: LruCache::with_hasher(
+            #[cfg(feature = "textlayout")]
+            shaping_run_cache: lru::LruCache::with_hasher(
                 std::num::NonZeroUsize::new(DEFAULT_LRU_CACHE_CAPACITY).unwrap(),
                 fnv_run,
             ),
-            shaped_words_cache: LruCache::with_hasher(
+            #[cfg(feature = "textlayout")]
+            shaped_words_cache: lru::LruCache::with_hasher(
                 std::num::NonZeroUsize::new(DEFAULT_LRU_CACHE_CAPACITY).unwrap(),
                 fnv_words,
             ),
@@ -312,14 +220,6 @@ impl Default for TextContextImpl {
 }
 
 impl TextContextImpl {
-    pub fn resize_shaping_run_cache(&mut self, capacity: std::num::NonZeroUsize) {
-        self.shaping_run_cache.resize(capacity);
-    }
-
-    pub fn resize_shaped_words_cache(&mut self, capacity: std::num::NonZeroUsize) {
-        self.shaped_words_cache.resize(capacity);
-    }
-
     pub fn add_font_dir<T: AsRef<FilePath>>(&mut self, path: T) -> Result<Vec<FontId>, ErrorKind> {
         let path = path.as_ref();
         let mut fonts = Vec::new();
@@ -389,6 +289,7 @@ impl TextContextImpl {
         self.fonts.get_mut(id.0)
     }
 
+    #[cfg(feature = "textlayout")]
     pub fn find_font<F, T>(&mut self, font_ids: &[Option<FontId>; 8], mut callback: F) -> Result<T, ErrorKind>
     where
         F: FnMut((FontId, &mut Font)) -> (bool, T),
@@ -427,56 +328,8 @@ impl TextContextImpl {
     }
 
     fn clear_caches(&mut self) {
+        #[cfg(feature = "textlayout")]
         self.shaped_words_cache.clear();
-    }
-
-    pub fn measure_text<S: AsRef<str>>(
-        &mut self,
-        x: f32,
-        y: f32,
-        text: S,
-        text_settings: &TextSettings,
-    ) -> Result<TextMetrics, ErrorKind> {
-        shape(x, y, self, text_settings, text.as_ref(), None)
-    }
-
-    pub fn break_text<S: AsRef<str>>(
-        &mut self,
-        max_width: f32,
-        text: S,
-        text_settings: &TextSettings,
-    ) -> Result<usize, ErrorKind> {
-        let layout = shape(0.0, 0.0, self, text_settings, text.as_ref(), Some(max_width))?;
-
-        Ok(layout.final_byte_index)
-    }
-
-    pub fn break_text_vec<S: AsRef<str>>(
-        &mut self,
-        max_width: f32,
-        text: S,
-        text_settings: &TextSettings,
-    ) -> Result<Vec<Range<usize>>, ErrorKind> {
-        let text = text.as_ref();
-
-        let mut res = Vec::new();
-        let mut start = 0;
-
-        while start < text.len() {
-            let Ok(index) = self.break_text(max_width, &text[start..], text_settings) else {
-                break;
-            };
-
-            if index == 0 {
-                break;
-            }
-
-            let index = start + index;
-            res.push(start..index);
-            start += &text[start..index].len();
-        }
-
-        Ok(res)
     }
 
     pub fn measure_font(&self, font_size: f32, font_ids: &[Option<FontId>; 8]) -> Result<FontMetrics, ErrorKind> {
@@ -488,374 +341,6 @@ impl TextContextImpl {
 
         Err(ErrorKind::NoFontFound)
     }
-}
-
-/// Represents the result of a text shaping run.
-#[derive(Clone, Default, Debug)]
-pub struct TextMetrics {
-    /// X-coordinate of the starting position for the shaped text.
-    pub x: f32,
-    /// Y-coordinate of the starting position for the shaped text.
-    pub y: f32,
-    width: f32,
-    height: f32,
-    /// Vector of shaped glyphs resulting from the text shaping run.
-    pub glyphs: Vec<ShapedGlyph>,
-    pub(crate) final_byte_index: usize,
-}
-
-impl TextMetrics {
-    pub(crate) fn scale(&mut self, scale: f32) {
-        self.x *= scale;
-        self.y *= scale;
-        self.width *= scale;
-        self.height *= scale;
-
-        for glyph in &mut self.glyphs {
-            glyph.x *= scale;
-            glyph.y *= scale;
-            glyph.width *= scale;
-            glyph.height *= scale;
-        }
-    }
-
-    /// width of the glyphs as drawn
-    pub fn width(&self) -> f32 {
-        self.width
-    }
-
-    /// height of the glyphs as drawn
-    pub fn height(&self) -> f32 {
-        self.height
-    }
-}
-
-// Shaper
-
-pub fn shape(
-    x: f32,
-    y: f32,
-    context: &mut TextContextImpl,
-    text_settings: &TextSettings,
-    text: &str,
-    max_width: Option<f32>,
-) -> Result<TextMetrics, ErrorKind> {
-    let id = ShapingId::new(text_settings.font_size, text_settings.font_ids, text, max_width);
-
-    if !context.shaping_run_cache.contains(&id) {
-        let metrics = shape_run(
-            context,
-            text_settings.font_size,
-            text_settings.font_ids,
-            text_settings.letter_spacing,
-            text,
-            max_width,
-        );
-        context.shaping_run_cache.put(id, metrics);
-    }
-
-    if let Some(mut metrics) = context.shaping_run_cache.get(&id).cloned() {
-        layout(x, y, context, &mut metrics, text_settings)?;
-
-        return Ok(metrics);
-    }
-
-    Err(ErrorKind::UnknownError)
-}
-
-fn shape_run(
-    context: &mut TextContextImpl,
-    font_size: f32,
-    font_ids: [Option<FontId>; 8],
-    letter_spacing: f32,
-    text: &str,
-    max_width: Option<f32>,
-) -> TextMetrics {
-    let mut result = TextMetrics {
-        x: 0.0,
-        y: 0.0,
-        width: 0.0,
-        height: 0.0,
-        glyphs: Vec::with_capacity(text.len()),
-        final_byte_index: 0,
-    };
-
-    let bidi_info = BidiInfo::new(text, Some(unicode_bidi::Level::ltr()));
-
-    // this controls whether we should break within words
-    let mut first_word_in_paragraph = true;
-
-    let Some(paragraph) = bidi_info.paragraphs.first() else {
-        return result;
-    };
-
-    let line = paragraph.range.clone();
-
-    let (levels, runs) = bidi_info.visual_runs(paragraph, line);
-
-    for run in runs {
-        let sub_text = &text[run.clone()];
-
-        if sub_text.is_empty() {
-            continue;
-        }
-
-        let hb_direction = if levels[run.start].is_rtl() {
-            rustybuzz::Direction::RightToLeft
-        } else {
-            rustybuzz::Direction::LeftToRight
-        };
-
-        let mut words = Vec::new();
-        let mut word_break_reached = false;
-        let mut byte_index = run.start;
-
-        for mut word_txt in sub_text.split_word_bounds() {
-            let id = ShapingId::new(font_size, font_ids, word_txt, max_width);
-
-            if !context.shaped_words_cache.contains(&id) {
-                let word = shape_word(word_txt, hb_direction, context, font_size, &font_ids, letter_spacing);
-                context.shaped_words_cache.put(id, word);
-            }
-
-            if let Some(Ok(word)) = context.shaped_words_cache.get(&id) {
-                let mut word = word.clone();
-
-                if let Some(max_width) = max_width {
-                    if result.width + word.width >= max_width {
-                        word_break_reached = true;
-                        if first_word_in_paragraph {
-                            // search for the largest prefix of the word that can fit
-                            let mut bytes_included = 0;
-                            let mut subword_width = 0.0;
-                            let target_width = max_width - result.width;
-                            for glyph in word.glyphs {
-                                bytes_included = glyph.byte_index;
-                                let glyph_width = glyph.advance_x + letter_spacing;
-
-                                // nuance: we want to include the first glyph even if it breaks
-                                // the bounds. this is to allow pathologically small bounds to
-                                // at least complete rendering
-                                if subword_width + glyph_width >= target_width && bytes_included != 0 {
-                                    break;
-                                }
-
-                                subword_width += glyph_width;
-                            }
-
-                            if bytes_included == 0 {
-                                // just in case - never mind!
-                                break;
-                            }
-
-                            let subword_txt = &word_txt[..bytes_included];
-                            let id = ShapingId::new(font_size, font_ids, subword_txt, Some(max_width));
-                            if !context.shaped_words_cache.contains(&id) {
-                                let subword = shape_word(
-                                    subword_txt,
-                                    hb_direction,
-                                    context,
-                                    font_size,
-                                    &font_ids,
-                                    letter_spacing,
-                                );
-                                context.shaped_words_cache.put(id, subword);
-                            }
-
-                            if let Some(Ok(subword)) = context.shaped_words_cache.get(&id) {
-                                // replace the outer variables so we can continue normally
-                                word = subword.clone();
-                                word_txt = subword_txt;
-                            } else {
-                                break;
-                            }
-                        } else if word.glyphs.iter().all(|g| g.c.is_whitespace()) {
-                            // the last word we've broken in the middle of is whitespace.
-                            // include this word for now, but we will discard its metrics in a moment.
-                        } else {
-                            // we are not breaking up words - discard this word
-                            break;
-                        }
-                    }
-                }
-
-                // if we have broken in the middle of whitespace, do not include this word in metrics
-                if !word_break_reached || !word.glyphs.iter().all(|g| g.c.is_whitespace()) {
-                    result.width += word.width;
-                }
-
-                for glyph in &mut word.glyphs {
-                    glyph.byte_index += byte_index;
-                    debug_assert!(text.get(glyph.byte_index..).is_some());
-                }
-                words.push(word);
-                first_word_in_paragraph = false;
-            }
-
-            byte_index += word_txt.len();
-
-            if word_break_reached {
-                break;
-            }
-        }
-
-        if levels[run.start].is_rtl() {
-            words.reverse();
-        }
-
-        for word in words {
-            result.glyphs.extend(word.glyphs.clone());
-        }
-
-        result.final_byte_index = byte_index;
-
-        if word_break_reached {
-            break;
-        }
-    }
-
-    result
-}
-
-fn shape_word(
-    word: &str,
-    hb_direction: rustybuzz::Direction,
-    context: &mut TextContextImpl,
-    font_size: f32,
-    font_ids: &[Option<FontId>; 8],
-    letter_spacing: f32,
-) -> Result<ShapedWord, ErrorKind> {
-    // find_font will call the closure with each font matching the provided style
-    // until a font capable of shaping the word is found
-    context.find_font(font_ids, |(font_id, font)| {
-        let face = font.face_ref();
-        // Call harfbuzz
-        let output = {
-            let mut buffer = rustybuzz::UnicodeBuffer::new();
-            buffer.push_str(word);
-            buffer.set_direction(hb_direction);
-
-            rustybuzz::shape(&face, &[], buffer)
-        };
-
-        let positions = output.glyph_positions();
-        let infos = output.glyph_infos();
-
-        let mut shaped_word = ShapedWord {
-            glyphs: Vec::with_capacity(positions.len()),
-            width: 0.0,
-        };
-
-        let mut has_missing = false;
-
-        for (position, (info, c)) in positions.iter().zip(infos.iter().zip(word.chars())) {
-            if info.glyph_id == 0 {
-                has_missing = true;
-            }
-
-            let scale = font.scale(font_size);
-
-            let mut g = ShapedGlyph {
-                x: 0.0,
-                y: 0.0,
-                c,
-                byte_index: info.cluster as usize,
-                font_id,
-                glyph_id: info
-                    .glyph_id
-                    .try_into()
-                    .expect("rustybuzz guarantees the output glyph id is u16"),
-                width: 0.0,
-                height: 0.0,
-                advance_x: position.x_advance as f32 * scale,
-                advance_y: position.y_advance as f32 * scale,
-                offset_x: position.x_offset as f32 * scale,
-                offset_y: position.y_offset as f32 * scale,
-                bearing_x: 0.0,
-                bearing_y: 0.0,
-            };
-
-            if let Some(glyph) = font.glyph(&face, g.glyph_id) {
-                g.width = glyph.metrics.width * scale;
-                g.height = glyph.metrics.height * scale;
-                g.bearing_x = glyph.metrics.bearing_x * scale;
-                g.bearing_y = glyph.metrics.bearing_y * scale;
-            }
-
-            shaped_word.width += g.advance_x + letter_spacing;
-            shaped_word.glyphs.push(g);
-        }
-
-        (has_missing, shaped_word)
-    })
-}
-
-// Calculates the x,y coordinates for each glyph based on their advances. Calculates total width and height of the shaped text run
-fn layout(
-    x: f32,
-    y: f32,
-    context: &mut TextContextImpl,
-    res: &mut TextMetrics,
-    text_settings: &TextSettings,
-) -> Result<(), ErrorKind> {
-    let mut cursor_x = x;
-    let mut cursor_y = y;
-
-    // Horizontal alignment
-    match text_settings.text_align {
-        Align::Center => cursor_x -= res.width / 2.0,
-        Align::Right => cursor_x -= res.width,
-        Align::Left => (),
-    }
-
-    res.x = cursor_x;
-
-    let mut min_y = cursor_y;
-    let mut max_y = cursor_y;
-
-    let mut ascender: f32 = 0.;
-    let mut descender: f32 = 0.;
-
-    for glyph in &mut res.glyphs {
-        let font = context.font_mut(glyph.font_id).ok_or(ErrorKind::NoFontFound)?;
-        let metrics = font.metrics(text_settings.font_size);
-        ascender = ascender.max(metrics.ascender());
-        descender = descender.min(metrics.descender());
-    }
-
-    let primary_metrics = context.find_font(&text_settings.font_ids, |(_, font)| {
-        (false, font.metrics(text_settings.font_size))
-    })?;
-    if ascender.abs() < f32::EPSILON {
-        ascender = primary_metrics.ascender();
-    }
-    if descender.abs() < f32::EPSILON {
-        descender = primary_metrics.descender();
-    }
-
-    // Baseline alignment
-    let alignment_offset_y = match text_settings.text_baseline {
-        Baseline::Top => ascender,
-        Baseline::Middle => (ascender + descender) / 2.0,
-        Baseline::Alphabetic => 0.0,
-        Baseline::Bottom => descender,
-    };
-
-    for glyph in &mut res.glyphs {
-        glyph.x = cursor_x + glyph.offset_x + glyph.bearing_x;
-        glyph.y = (cursor_y + alignment_offset_y).round() + glyph.offset_y - glyph.bearing_y;
-
-        min_y = min_y.min(glyph.y);
-        max_y = max_y.max(glyph.y + glyph.height);
-
-        cursor_x += glyph.advance_x + text_settings.letter_spacing;
-        cursor_y += glyph.advance_y;
-    }
-
-    res.y = min_y;
-    res.height = max_y - min_y;
-
-    Ok(())
 }
 
 // Renderer
@@ -911,7 +396,7 @@ impl GlyphAtlas {
         canvas: &mut Canvas<T>,
         font_id: FontId,
         font: &Font,
-        font_face: &rustybuzz::Face<'_>,
+        font_face: &ttf_parser::Face<'_>,
         glyphs: impl Iterator<Item = PositionedGlyph>,
         font_size: f32,
         line_width: f32,
@@ -1005,7 +490,7 @@ impl GlyphAtlas {
         line_width: f32,
         mode: RenderMode,
         font: &Font,
-        font_face: &rustybuzz::Face<'_>,
+        font_face: &ttf_parser::Face<'_>,
         glyph_id: u16,
     ) -> Result<Option<RenderedGlyph>, ErrorKind> {
         let padding = GLYPH_PADDING + GLYPH_MARGIN;
